+++ example/lib/pages/radio_page.dart
@@ -53,7 +53,7 @@ class _RadioPageState extends State<RadioPage> {
               const SizedBox(height: 10),
             ],
             const Divider(),
-            for (var i = 0; i < 3; ++i)
+            for (var i = 0; i < 3; ++i) ...[
               YaruRadioListTile<int>(
                 value: i,
                 groupValue: _listTileValue,
@@ -61,6 +61,7 @@ class _RadioPageState extends State<RadioPage> {
                 toggleable: true,
                 title: const Text('YaruRadioListTile'),
               ),
+            ],
           ],
         );
       },