+++ lib/src/widgets/yaru_checkbox_list_tile.dart
@@ -13,7 +13,7 @@ import 'package:yaru/yaru.dart';
 ///  * [YaruRadioListTile], a similar widget for radio buttons.
 ///  * [YaruSwitchListTile], a similar widget for switches.
 ///  * [ListTile] and [YaruCheckbox], the widgets from which this widget is made.
-class YaruCheckboxListTile extends StatelessWidget {
+class YaruCheckboxListTile extends YaruToggleListTile {
   /// Creates a combination of a [ListTile] and a [YaruCheckbox].
   ///
   /// See [CheckboxListTile].
@@ -21,26 +21,20 @@ class YaruCheckboxListTile extends StatelessWidget {
     super.key,
     required this.value,
     required this.onChanged,
-    this.tileColor,
-    this.title,
-    this.subtitle,
-    this.isThreeLine = false,
-    this.dense,
-    this.secondary,
-    this.selected = false,
-    this.controlAffinity = ListTileControlAffinity.platform,
-    this.autofocus = false,
-    this.contentPadding,
     this.tristate = false,
-    this.shape,
-    this.selectedTileColor,
-    this.visualDensity,
-    this.focusNode,
-    this.enableFeedback,
-    this.mouseCursor,
-    this.hasFocusBorder,
-  }) : assert(tristate || value != null),
-       assert(!isThreeLine || subtitle != null);
+    super.control,
+    super.title,
+    super.subtitle,
+    super.secondary,
+    super.controlAffinity,
+    super.autofocus = false,
+    super.shape,
+    super.focusNode,
+    super.enableFeedback,
+    super.mouseCursor,
+    super.hasFocusBorder,
+    super.contentPadding,
+  }) : assert(tristate || value != null);
 
   /// See [CheckboxListTile.value].
   final bool? value;
@@ -48,59 +42,8 @@ class YaruCheckboxListTile extends StatelessWidget {
   /// See [CheckboxListTile.onChanged].
   final ValueChanged<bool?>? onChanged;
 
-  /// See [CheckboxListTile.tileColor].
-  final Color? tileColor;
-
-  /// See [CheckboxListTile.title].
-  final Widget? title;
-
-  /// See [CheckboxListTile.subtitle].
-  final Widget? subtitle;
-
-  /// See [CheckboxListTile.secondary].
-  final Widget? secondary;
-
-  /// See [CheckboxListTile.isThreeLine].
-  final bool isThreeLine;
-
-  /// See [CheckboxListTile.dense].
-  final bool? dense;
-
-  /// See [CheckboxListTile.selected].
-  final bool selected;
-
-  /// See [CheckboxListTile.controlAffinity].
-  final ListTileControlAffinity controlAffinity;
-
-  /// See [CheckboxListTile.autofocus].
-  final bool autofocus;
-
-  /// See [CheckboxListTile.contentPadding].
-  final EdgeInsetsGeometry? contentPadding;
-
   /// See [CheckboxListTile.tristate].
   final bool tristate;
-
-  /// See [CheckboxListTile.shape].
-  final ShapeBorder? shape;
-
-  /// See [CheckboxListTile.selectedTileColor].
-  final Color? selectedTileColor;
-
-  /// See [CheckboxListTile.visualDensity].
-  final VisualDensity? visualDensity;
-
-  /// See [CheckboxListTile.focusNode].
-  final FocusNode? focusNode;
-
-  /// See [CheckboxListTile.enableFeedback].
-  final bool? enableFeedback;
-
-  /// See [CheckboxListTile.mouseCursor].
-  final MouseCursor? mouseCursor;
-
-  /// Whether to display the default focus border on focus or not.
-  final bool? hasFocusBorder;
 
   void _handleValueChange() {
     assert(onChanged != null);
@@ -118,54 +61,78 @@ class YaruCheckboxListTile extends StatelessWidget {
   }
 
   @override
+  State<StatefulWidget> createState() => _YaruCheckboxListTileState();
+}
+
+class _YaruCheckboxListTileState extends State<YaruCheckboxListTile> {
+  bool _tileHasFocus = false;
+  late final FocusNode _focusNode;
+
+  @override
+  void initState() {
+    super.initState();
+    _focusNode = widget.focusNode ?? FocusNode();
+  }
+
+  @override
+  void dispose() {
+    super.dispose();
+    if (widget.focusNode == null) {
+      _focusNode.dispose();
+    }
+  }
+
+  @override
   Widget build(BuildContext context) {
     Widget? leading, trailing;
-    final Widget control = YaruCheckbox(
-      value: value,
-      onChanged: onChanged,
-      autofocus: autofocus,
-      tristate: tristate,
-      mouseCursor: mouseCursor,
-      hasFocusBorder: false,
-    );
+    final control =
+        widget.control ??
+        YaruCheckbox(
+          value: widget.value,
+          onChanged: widget.onChanged,
+          autofocus: widget.autofocus,
+          tristate: widget.tristate,
+          mouseCursor: widget.mouseCursor,
+        );
 
-    switch (controlAffinity) {
+    switch (widget.controlAffinity) {
       case ListTileControlAffinity.leading:
+      case ListTileControlAffinity.platform:
         leading = control;
-        trailing = secondary;
+        trailing = widget.secondary;
         break;
       case ListTileControlAffinity.trailing:
-      case ListTileControlAffinity.platform:
-        leading = secondary;
+        leading = widget.secondary;
         trailing = control;
         break;
     }
 
-    final tile = ListTile(
+    final tile = YaruListTile(
       leading: leading,
-      title: title,
-      subtitle: subtitle,
+      title: widget.title,
+      subtitle: widget.subtitle,
       trailing: trailing,
-      isThreeLine: isThreeLine,
-      dense: dense,
-      enabled: onChanged != null,
-      onTap: onChanged != null ? _handleValueChange : null,
-      selected: selected,
-      autofocus: autofocus,
-      contentPadding: contentPadding,
-      shape: shape,
-      selectedTileColor: selectedTileColor,
-      tileColor: tileColor,
-      visualDensity: visualDensity,
-      focusNode: focusNode,
-      enableFeedback: enableFeedback,
-      mouseCursor: mouseCursor,
+      enabled: widget.onChanged != null,
+      onTap: widget.onChanged != null ? widget._handleValueChange : null,
+      autofocus: widget.autofocus,
+      customBorder: widget.shape,
+      focusNode: _focusNode,
+      enableFeedback: widget.enableFeedback,
+      mouseCursor: widget.mouseCursor,
+      contentPadding: widget.contentPadding,
+      hasFocusBorder: false,
+      onFocusChange: (focus) => setState(() {
+        _tileHasFocus = _focusNode.hasPrimaryFocus;
+      }),
     );
 
     return MergeSemantics(
-      child: hasFocusBorder ?? YaruTheme.maybeOf(context)?.focusBorders == true
+      child:
+          widget.hasFocusBorder ??
+              YaruTheme.maybeOf(context)?.focusBorders == true
           ? YaruFocusBorder.primary(
               borderStrokeAlign: BorderSide.strokeAlignInside,
+              borderColor: _tileHasFocus ? null : Colors.transparent,
               child: tile,
             )
           : tile,