+++ lib/src/widgets/yaru_radio_list_tile.dart
@@ -15,7 +15,7 @@ import 'package:yaru/yaru.dart';
 ///  * [YaruCheckboxListTile], a similar widget for checkboxes.
 ///  * [YaruSwitchListTile], a similar widget for switches.
 ///  * [ListTile] and [YaruRadio], the widgets from which this widget is made.
-class YaruRadioListTile<T> extends StatelessWidget {
+class YaruRadioListTile<T> extends YaruToggleListTile {
   /// Creates a combination of a [ListTile] and a [YaruRadio].
   ///
   /// See [RadioListTile].
@@ -25,24 +25,18 @@ class YaruRadioListTile<T> extends StatelessWidget {
     required this.groupValue,
     required this.onChanged,
     this.toggleable = false,
-    this.title,
-    this.subtitle,
-    this.isThreeLine = false,
-    this.dense,
-    this.secondary,
-    this.selected = false,
-    this.controlAffinity = ListTileControlAffinity.platform,
-    this.autofocus = false,
-    this.contentPadding,
-    this.shape,
-    this.tileColor,
-    this.selectedTileColor,
-    this.visualDensity,
-    this.focusNode,
-    this.enableFeedback,
-    this.mouseCursor,
-    this.hasFocusBorder,
-  }) : assert(!isThreeLine || subtitle != null);
+    super.control,
+    super.title,
+    super.subtitle,
+    super.secondary,
+    super.controlAffinity,
+    super.autofocus = false,
+    super.shape,
+    super.focusNode,
+    super.mouseCursor,
+    super.hasFocusBorder,
+    super.contentPadding,
+  });
 
   /// See [RadioListTile.value].
   final T value;
@@ -56,57 +50,6 @@ class YaruRadioListTile<T> extends StatelessWidget {
   /// See [RadioListTile.toggleable].
   final bool toggleable;
 
-  /// See [RadioListTile.title].
-  final Widget? title;
-
-  /// See [RadioListTile.subtitle].
-  final Widget? subtitle;
-
-  /// See [RadioListTile.secondary].
-  final Widget? secondary;
-
-  /// See [RadioListTile.isThreeLine].
-  final bool isThreeLine;
-
-  /// See [RadioListTile.dense].
-  final bool? dense;
-
-  /// See [RadioListTile.selected].
-  final bool selected;
-
-  /// See [RadioListTile.controlAffinity].
-  final ListTileControlAffinity controlAffinity;
-
-  /// See [RadioListTile.autofocus].
-  final bool autofocus;
-
-  /// See [RadioListTile.contentPadding].
-  final EdgeInsetsGeometry? contentPadding;
-
-  /// See [RadioListTile.shape].
-  final ShapeBorder? shape;
-
-  /// See [RadioListTile.tileColor].
-  final Color? tileColor;
-
-  /// See [RadioListTile.selectedTileColor].
-  final Color? selectedTileColor;
-
-  /// See [RadioListTile.visualDensity].
-  final VisualDensity? visualDensity;
-
-  /// See [RadioListTile.focusNode].
-  final FocusNode? focusNode;
-
-  /// See [RadioListTile.enableFeedback].
-  final bool? enableFeedback;
-
-  /// See [RadioListTile.mouseCursor].
-  final MouseCursor? mouseCursor;
-
-  /// Whether to display the default focus border on focus or not.
-  final bool? hasFocusBorder;
-
   void _handleValueChange() {
     assert(onChanged != null);
     if (groupValue != value || !toggleable) {
@@ -117,55 +60,79 @@ class YaruRadioListTile<T> extends StatelessWidget {
   }
 
   @override
+  State<StatefulWidget> createState() => _YaruRadioListTileState<T>();
+}
+
+class _YaruRadioListTileState<T> extends State<YaruRadioListTile<T>> {
+  bool _tileHasFocus = false;
+  late final FocusNode _focusNode;
+
+  @override
+  void initState() {
+    super.initState();
+    _focusNode = widget.focusNode ?? FocusNode();
+  }
+
+  @override
+  void dispose() {
+    super.dispose();
+    if (widget.focusNode == null) {
+      _focusNode.dispose();
+    }
+  }
+
+  @override
   Widget build(BuildContext context) {
     Widget? leading, trailing;
-    final Widget control = YaruRadio<T>(
-      value: value,
-      groupValue: groupValue,
-      onChanged: onChanged,
-      toggleable: toggleable,
-      autofocus: autofocus,
-      mouseCursor: mouseCursor,
-      hasFocusBorder: false,
-    );
+    final control =
+        widget.control ??
+        YaruRadio<T>(
+          value: widget.value,
+          groupValue: widget.groupValue,
+          onChanged: widget.onChanged,
+          toggleable: widget.toggleable,
+          autofocus: widget.autofocus,
+          mouseCursor: widget.mouseCursor,
+        );
 
-    switch (controlAffinity) {
+    switch (widget.controlAffinity) {
       case ListTileControlAffinity.leading:
       case ListTileControlAffinity.platform:
         leading = control;
-        trailing = secondary;
+        trailing = widget.secondary;
         break;
       case ListTileControlAffinity.trailing:
-        leading = secondary;
+        leading = widget.secondary;
         trailing = control;
         break;
     }
 
-    final tile = ListTile(
+    final tile = YaruListTile(
       leading: leading,
-      title: title,
-      subtitle: subtitle,
+      title: widget.title,
+      subtitle: widget.subtitle,
       trailing: trailing,
-      isThreeLine: isThreeLine,
-      dense: dense,
-      enabled: onChanged != null,
-      shape: shape,
-      tileColor: tileColor,
-      selectedTileColor: selectedTileColor,
-      onTap: onChanged != null ? _handleValueChange : null,
-      selected: selected,
-      autofocus: autofocus,
-      contentPadding: contentPadding,
-      visualDensity: visualDensity,
-      focusNode: focusNode,
-      enableFeedback: enableFeedback,
-      mouseCursor: mouseCursor,
+      enabled: widget.onChanged != null,
+      onTap: widget.onChanged != null ? widget._handleValueChange : null,
+      autofocus: widget.autofocus,
+      focusNode: _focusNode,
+      hoverColor: widget.hoverColor,
+      mouseCursor: widget.mouseCursor,
+      customBorder: widget.shape,
+      contentPadding: widget.contentPadding,
+      hasFocusBorder: false,
+      onFocusChange: (focus) => setState(() {
+        _tileHasFocus = _focusNode.hasPrimaryFocus;
+      }),
     );
 
     return MergeSemantics(
-      child: hasFocusBorder ?? YaruTheme.maybeOf(context)?.focusBorders == true
+      child:
+          widget.hasFocusBorder ??
+              YaruTheme.maybeOf(context)?.focusBorders == true
           ? YaruFocusBorder.primary(
               borderStrokeAlign: BorderSide.strokeAlignInside,
+              borderColor: _tileHasFocus ? null : Colors.transparent,
               child: tile,
             )
           : tile,